Refactor ActiveAdmin AdminLinkTo helpers to a proper module
And include it properly
